+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Add load more threshold for `MessageListView` and `streamLoadMoreThreshold` attribute
 - Fix handling of the `streamShowReadState` attribute on `MessageListView`
 - Add `streamShowDeliveredState` XML attribute to `MessageListView`
+- Add "loading more" indicator to the `MessageListView`
 
 ## stream-chat-android-client
 - Fix guest user authentication